# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SessionInput {
  role: string;
  difficulty: string;
  totalQuestions: number;
}

export interface Feedback {
  score: number;
  strengths: string[];
  improvements: string[];
  modelAnswer: string;
  keywordCoverage: number;
}

export interface Session {
  id: string;
  role: string;
  difficulty: string;
  totalQuestions: number;
  questions?: string[];
  answers?: string[];
  scores?: number[];
  feedbackList?: Feedback[];
  createdAt: string;
}

export interface QuestionInput {
  role: string;
  difficulty: string;
}

export interface Question {
  question: string;
  hint: string;
  expectedKeywords: string[];
}

export interface EvaluationInput {
  sessionId: string;
  questionIndex: number;
  question: string;
  answer: string;
  expectedKeywords: string[];
}

export interface JdInput {
  jdText: string;
}

export interface InterviewStats {
  totalSessions: number;
  averageScore: number;
  topRole: string;
  totalQuestions: number;
}