# lib/api-zod/src/generated/types/evaluationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EvaluationInput {
  sessionId: string;
  questionIndex: number;
  question: string;
  answer: string;
  expectedKeywords: string[];
}