# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  score: number;
  strengths: string[];
  improvements: string[];
  modelAnswer: string;
  keywordCoverage: number;
}