# lib/api-zod/src/generated/types/interviewStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InterviewStats {
  totalSessions: number;
  averageScore: number;
  topRole: string;
  totalQuestions: number;
}