# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Feedback } from './feedback';

export interface Session {
  id: string;
  role: string;
  difficulty: string;
  totalQuestions: number;
  questions?: string[];
  answers?: string[];
  scores?: number[];
  feedbackList?: Feedback[];
  createdAt: string;
}